Optimized & Restructured ETC Mini App
